+++ test/test_docs_flask.py
@@ -21,25 +21,25 @@ class TestDocsFlask(TestCase):
         self.harness.teardown_app(mock.sentinel.sphinx_app)
 
     def test_harness_iter_annotations(self):
-        self.assertEqual(len(self.annotations), 4)
+        assert len(self.annotations) == 4
 
         heading, rule, view_class, annotations = self.annotations[0]
-        self.assertEqual(heading, 'API Status')
-        self.assertEqual(rule.rule, '/')
-        self.assertEqual([annotation.http_method for annotation in annotations],
-                         ['GET'])
+        assert heading == 'API Status'
+        assert rule.rule == '/'
+        assert [annotation.http_method
+                for annotation in annotations] == ['GET']
 
         heading, rule, view_class, annotations = self.annotations[1]
-        self.assertEqual(heading, 'Notes (v1)')
-        self.assertEqual(rule.rule, '/note/')
-        self.assertEqual([annotation.http_method for annotation in annotations],
-                         ['GET', 'POST'])
+        assert heading == 'Notes (v1)'
+        assert rule.rule == '/note/'
+        assert ([annotation.http_method for annotation in annotations] ==
+                ['GET', 'POST'])
 
         heading, rule, view_class, annotations = self.annotations[2]
-        self.assertEqual(heading, 'Notes (v1)')
-        self.assertEqual(rule.rule, '/note/<int:note_id>/')
-        self.assertEqual([annotation.http_method for annotation in annotations],
-                         ['GET', 'PUT', 'DELETE'])
+        assert heading, 'Notes (v1)'
+        assert rule.rule == '/note/<int:note_id>/'
+        assert ([annotation.http_method for annotation in annotations] ==
+                ['GET', 'PUT', 'DELETE'])
 
         heading, rule, view_class, annotations = self.annotations[3]
         assert 'Z' == heading
@@ -51,13 +51,14 @@ class TestDocsFlask(TestCase):
         annotation = annotations[0]
         result = self.harness.request(rule, view_class, annotation)
         result['response'] = json.loads(result['response'])
-        self.assertEqual(result, {
+        assert result == {
             'method': 'GET',
             'params': {},
             'response': {'body': 'Example body',
                          'done': True,
                          'note_id': 1},
-            'url': 'http://127.0.0.1/note/1/?note_type=quick'})
+            'url': 'http://127.0.0.1/note/1/?note_type=quick',
+        }
 
     def test_harness_request_get_list_and_dict_params(self):
         """
@@ -68,30 +69,29 @@ class TestDocsFlask(TestCase):
         annotation = annotations[0]
         result = self.harness.request(rule, view_class, annotation)
         result['response'] = json.loads(result['response'])
-        expected = {
-
+        assert result == {
             'method': 'GET',
             'params': {},
             'response': {},
             'url': ('http://127.0.0.1/example/list-obj/?note_types=%5B%22quick'
                     '%22%5D&a_note=%7B%22body%22%3A+%22Example+Body%22%2C+%22'
-                    'done%22%3A+true%2C+%22note_id%22%3A+1%7D')
+                    'done%22%3A+true%2C+%22note_id%22%3A+1%7D'),
         }
-        self.assertEqual(expected, result)
 
     def test_harness_request_post(self):
         _, rule, view_class, annotations = self.annotations[1]
         annotation = annotations[1]
         result = self.harness.request(rule, view_class, annotation)
         result['response'] = json.loads(result['response'])
-        self.assertEqual(result, {
+        assert result == {
             'method': 'POST',
             'params': {'body': 'body',
                        'done': False},
             'response': {'body': 'body',
                          'done': False,
                          'note_id': 2},
-            'url': 'http://127.0.0.1/note/'})
+            'url': 'http://127.0.0.1/note/',
+        }
 
     def test_harness_request_delete(self):
         response_mock = mock.Mock(data='')
@@ -100,30 +100,29 @@ class TestDocsFlask(TestCase):
         annotation = annotations[2]
         result = self.harness.request(rule, view_class, annotation)
 
-        expected_url = '/note/1/'
         self.harness.test_client.delete.assert_called_with(
-            expected_url, data={}, headers={})
-        expected_result = {
+            '/note/1/', data={}, headers={})
+        assert result == {
             'url': 'http://127.0.0.1/note/1/',
             'params': {},
             'method': 'DELETE',
             'response': ''
         }
-        self.assertEqual(expected_result, result)
 
     def test_harness_define_example_values(self):
         _, rule, view_class, annotations = self.annotations[1]
         annotation = annotations[1]
         result = self.harness.request(rule, view_class, annotation)
         result['response'] = json.loads(result['response'])
-        self.assertEqual(result, {
+        assert result == {
             'method': 'POST',
             'params': {'body': 'body',
                        'done': False},
             'response': {'body': 'body',
                          'done': False,
                          'note_id': 2},
-            'url': 'http://127.0.0.1/note/'})
+            'url': 'http://127.0.0.1/note/',
+        }
 
         # Now override the values and validate that they're different
         self.harness.define_example_values('POST', '/note/', {
@@ -131,7 +130,7 @@ class TestDocsFlask(TestCase):
         })
         result = self.harness.request(rule, view_class, annotation)
         result['response'] = json.loads(result['response'])
-        self.assertEqual(result['params'], {'body': 'This is a replaced body.'})
+        assert result['params'] == {'body': 'This is a replaced body.'}
 
         # Override, but with update=True, which should merge the values
         self.harness.define_example_values('POST', '/note/', {
@@ -139,5 +138,5 @@ class TestDocsFlask(TestCase):
         }, update=True)
         result = self.harness.request(rule, view_class, annotation)
         result['response'] = json.loads(result['response'])
-        self.assertEqual(result['params'], {'body': 'This is an updated body.',
-                                            'done': False})
+        assert result['params'] == {'body': 'This is an updated body.',
+                                    'done': False}