+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Next release (in development)
 -----------------------------
 
 * Added new `UnionType` to types that allows a value to be one of `n` types.
+* Don't filter out request parameters not defined in the type object for routes
+  that specify a `req_obj_type`.
 
 v3.8.2 (2018-07-02)
 -------------------