+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ Release History
 Next release (in development)
 -----------------------------
 
+v3.9.0 (2018-07-13)
+-------------------
+
 * Added new `UnionType` to types that allows a value to be one of `n` types.
 * Don't filter out request parameters not defined in the type object for routes
   that specify a `req_obj_type`.